fix: make syncDir a no-op on Windows
writeState's directory fsync panicked on Windows with ERROR_ACCESS_DENIED
because os.Open opens a directory handle read-only, and FlushFileBuffers
(the syscall behind f.Sync) requires GENERIC_WRITE. This took down weed
master on every term-persisting write — see seaweedfs/seaweedfs#9292.

Split syncDir into build-tagged files: keep the os.Open + f.Sync path on
unix, and return nil on Windows. Directory fsync has no portable
equivalent there anyway — NTFS journals rename metadata, so the rename
of state.tmp -> state is durable through the filesystem's own ordering.
